pyproject.toml
[project]
name = "camera-perception"
version = "0.1.0"
description = "Camera-based perception (traversability + detection) for off-road autonomous mobility"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "numpy>=1.26",
    "pillow>=10.0",
    "opencv-python>=4.9",
    "albumentations>=1.4",
    "pyyaml>=6.0",
    "omegaconf>=2.3",
    "hydra-core>=1.3",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "lightning>=2.2",
    "torchmetrics>=1.4",
    "transformers>=4.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipykernel>=6.29",
    "jupyter>=1.0",
]
logging = [
    "wandb>=0.17",
    "tensorboard>=2.16",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/camera_perception"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
